Update page top spacing to the site class
The page class is only found in markup for the page template. This
moves the top spacing to a new site style that was added to the
default template layout. This prevents it from being applied twice,
one globally on the default template layout and once in page
template layouts.